Moved parseTimeValue to time/index.ts
