+++ server/src/util.ts
@@ -1,10 +1,7 @@
 import { Diagnostic, DiagnosticSeverity, Position, Range } from "vscode-languageserver";
 import { Section } from "./configTree/section";
-import { dateError } from "./messageUtil";
 import { settingsMap } from "./resources";
 import { Setting } from "./setting";
-import { TimeParseError } from "./time/timeParseError";
-import { TimeParser } from "./time/timeParser";
 
 const DIAGNOSTIC_SOURCE: string = "Axibase Charts";
 
@@ -196,43 +193,6 @@ export function createRange(start: number, length: number, lineNumber: number) {
         Position.create(lineNumber, start),
         Position.create(lineNumber, start + length),
     );
-}
-
-const timeParseCache: Map<Setting, Date> = new Map<Setting, Date>();
-/**
- * Parses value of time setting, adds diagnostic to `errors` if any error during parsing was thrown.
- *
- * @param timeSetting - Date setting, which value is need to be parsed
- * @param timeParser - Util class, containig methods for date parsing
- * @param errors - Array of diagnosics, to which information about error is added
- * @returns Value of `timeSetting`, parsed to Date.
- */
-export function parseTimeValue(timeSetting: Setting, section: Section, errors: Diagnostic[]): Date {
-    let parsedValue;
-    if (timeSetting != null) {
-        if (timeParseCache.has(timeSetting)) {
-            const cached = timeParseCache.get(timeSetting);
-            if (cached instanceof Date) {
-                return cached;
-            }
-            return null;
-        }
-        try {
-            const timeZoneValue = getValueOfSetting("time-zone", section);
-            const timeParser = new TimeParser(timeZoneValue as string);
-            parsedValue = timeParser.parseDateTemplate(timeSetting.value);
-            timeParseCache.set(timeSetting, parsedValue);
-        } catch (err) {
-            if (err instanceof TimeParseError) {
-                const diagnostic = createDiagnostic(timeSetting.textRange,
-                    dateError(err.message, timeSetting.displayName));
-                errors.push(diagnostic);
-            } else {
-                throw err;
-            }
-        }
-    }
-    return parsedValue;
 }
 
 /**
